feat: enhance job search functionality with API integration and improved UI components

+++ services/ai/app/schemas/company_trust.py
@@ -14,7 +14,7 @@ class TrustReport(BaseModel):
 
     trust_score: int = 0
 
-    confidence: int = 0
+    confidence: float = 0
 
     breakdown: TrustBreakdown
